pyproject.toml
[project]
name = "wawa-fg"
version = "0.1.0"
description = "Feature Gating 管理系统 - 基于业务字段表达式的功能控制平台"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi[standard]>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "motor>=3.3.2",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt==4.3.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "cachetools>=5.3.2",
    "pydantic-settings>=2.12.0",
    "itsdangerous>=2.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
